requirements.txt
certifi
click
colorama
cycler
Flask
fonttools
importlib-metadata
itsdangerous
Jinja2
joblib
kiwisolver
kneed
MarkupSafe
matplotlib
mysql-connector-python
numpy
packaging
pandas
Pillow
protobuf
pyparsing
python-dateutil
pytz
scikit-learn
scipy
six
sklearn
threadpoolctl
Werkzeug
wincertstore
zipp

